+++ components/modules/Home/includes/js/map.js
@@ -97,6 +97,9 @@
       };
       map.update_events();
       return cs.home.delete_event = function(id) {
+        if (!confirm('Точно видалити?')) {
+          return;
+        }
         $.ajax({
           url: "api/Home/events/" + id,
           type: 'delete',
